Move encoders inside test folder
