+++ src/domain/contracts/__tests__/encoders/safe-encoder.builder.ts
